move notification handling to clients

+++ services/server/core/models/user.go
@@ -6,6 +6,6 @@ type User struct {
 	Email   string `json:"email" dapper:"email"`
 	Name    string `json:"name" dapper:"name"`
 	Avatar  string `json:"picture,omitempty" dapper:"avatar"`
-	Setting int32  `dapper:"settings"`
+	Setting int64  `dapper:"settings"`
 	Role    string `json:"role" dapper:"role"`
 }
